convert js to ts framework files
